@@ -1,4 +1,5 @@
 import * as axios from "axios";
+// import FormData from 'form-data'
 
 const instance = axios.create({
     withCredentials: true,
@@ -34,7 +35,12 @@ export const profileAPI = {
     savePhoto(image) {
         let formData = new FormData()
         formData.append('image', image)
-        return instance.put('profile/photo', formData, {headers: {'Content-Type': 'multipart/form-data'}})
+        for (var key of formData.entries()) {
+            console.log(key[0] + ', ' + key[1]);
+        }
+
+        debugger
+        return instance.post('profile/photo', formData, {headers: {'Content-Type': 'multipart/form-data'}})
     }
 }
 
