@@ -1,7 +1,7 @@
-import {Field, reduxForm} from "redux-form";
-import React from "react";
-import {Textarea} from "../common/FormsControls/FormsControl";
-import {maxLengthCreator, required} from "../../utils/validators/validator";
+import {Field, reduxForm} from "redux-form"
+import React from "react"
+import {Textarea} from "../common/FormsControls/FormsControl"
+import {maxLengthCreator, required} from "../../utils/validators/validator"
 
 let maxLength50 = maxLengthCreator(50)
 
