@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
-import s from './ProfileHeader.module.css';
+import s from './ProfileInfo.module.scss';
 
-const ProfileStatusWithHooks = (props) => {
+const ProfileStatus = (props) => {
     let [editMode, setEditMode] = useState(false)
     let [status, setStatus] = useState(props.status)
 
@@ -15,7 +15,7 @@ const ProfileStatusWithHooks = (props) => {
 
     const deactivateEditMode = () => {
         setEditMode(false)
-        props.updateStatus(status)
+        props.updateStatusTC(status)
     }
 
     const onStatusChange = (e) => {
@@ -40,4 +40,4 @@ const ProfileStatusWithHooks = (props) => {
     )
 }
 
-export default ProfileStatusWithHooks;
+export default ProfileStatus;
