@@ -1,55 +1,50 @@
-import React from 'react';
-import {connect} from "react-redux";
-import {compose} from "redux";
-import Users from "./Users";
-import Preloader from "../common/Preloader/Preloader";
+ import React, {useEffect} from 'react'
+import {connect} from "react-redux"
+
 import {
-    followTC as follow,
-    unfollowTC as unfollow,
-    setCurrentPageAC as setCurrentPage,
-    toggleIsFollowingInProgressAC as toggleIsFollowingInProgress,
-    requestUsersTC as requestUsers
-} from "../../redux/users-reducer";
+    followTC, unfollowTC,
+    setCurrentPageAC, requestUsersTC,
+    toggleIsFollowingInProgressAC
+} from "../../redux/users-reducer"
 import {
-    getUsers,
-    getCurrentPage,
-    getIsFetching,
-    getIsFollowingInProgress,
-    getPageSize,
-    getTotalUsersCount
-} from "../../redux/users-selectors";
+    getUsers, getCurrentPage,
+    getIsFetching, getIsFollowingInProgress,
+    getPageSize, getTotalUsersCount
+} from "../../redux/users-selectors"
 
+import Users from "./Users"
+import Preloader from "../common/Preloader/Preloader"
 
+const UsersContainer = (props) => {
+    const {totalUsersCount, pageSize, users, currentPage, isFetching, followTC, unfollowTC, isFollowingInProgress} = props
 
+    useEffect(() => {
+        props.requestUsersTC(currentPage, pageSize)
+    }, [])
 
-class UsersContainer extends React.Component {
-    componentDidMount() {
-        const {currentPage, pageSize} = this.props
-        this.props.requestUsers(currentPage, pageSize)
+    const onPageChanged = (pageNumber) => {
+        const {pageSize} = props
+        props.setCurrentPageAC(pageNumber)
+        props.requestUsersTC(pageNumber, pageSize)
     }
 
-    onPageChanged = (pageNumber) => {
-        const {pageSize} = this.props
-        this.props.setCurrentPage(pageNumber)
-        this.props.requestUsers(pageNumber, pageSize)
-    }
-
-    render() {
-        const {totalUsersCount, pageSize, users, currentPage, isFetching, follow, unfollow, isFollowingInProgress} = this.props
-        return <>
-            {this.props.isFetching ? <Preloader/> : null}
+    return (
+        <>
+            {props.isFetching ? <Preloader/> : null}
             <Users totalUsersCount={totalUsersCount}
                    pageSize={pageSize}
                    users={users}
                    currentPage={currentPage}
                    isFetching={isFetching}
-                   onPageChanged={this.onPageChanged}
-                   follow={follow}
-                   unfollow={unfollow}
+                   onPageChanged={onPageChanged}
+                   followTC={followTC}
+                   unfollowTC={unfollowTC}
                    isFollowingInProgress={isFollowingInProgress}
             />
         </>
-    }
+    )
+
+
 }
 
 let mapStateToProps = (state) => ({
@@ -61,13 +56,9 @@ let mapStateToProps = (state) => ({
     isFollowingInProgress: getIsFollowingInProgress(state)
 })
 
-let mapDispatchToPropsObj = {
-    follow, unfollow, setCurrentPage,
-    toggleIsFollowingInProgress, requestUsers
+let mapDispatchToProps = {
+    followTC, unfollowTC, setCurrentPageAC,
+    toggleIsFollowingInProgressAC, requestUsersTC
 }
 
-const UsersContainerC = compose(
-    connect(mapStateToProps, mapDispatchToPropsObj),
-)(UsersContainer)
-
-export default UsersContainerC
+export default connect(mapStateToProps, mapDispatchToProps)(UsersContainer)
