@@ -1,9 +1,9 @@
 import React, {Suspense} from "react";
 import Preloader from "../common/Preloader/Preloader";
 
-let mapStateToPropsForRedirect = (state) => ({
-    isAuth: state.auth.isAuth
-})
+// let mapStateToPropsForRedirect = (state) => ({
+//     isAuth: state.auth.isAuth
+// })
 
 export const withSuspense = (Component) => (props) => (
     <Suspense fallback={<Preloader />}>
