@@ -1,11 +1,12 @@
-import {applyMiddleware, combineReducers, createStore, compose} from "redux";
-import profileReducer from "./profile-reducer";
-import messagesReducer from "./messages-reducer";
-import usersReducer from "./users-reducer";
-import appReducer from "./app-reducer";
-import authReducer from "./auth-reducer";
+import {applyMiddleware, combineReducers, createStore, compose} from "redux"
 import thunkMiddleware from "redux-thunk"
 import {reducer as formReducer} from 'redux-form'
+
+import profileReducer from "./profile-reducer"
+import messagesReducer from "./messages-reducer"
+import usersReducer from "./users-reducer"
+import appReducer from "./app-reducer"
+import authReducer from "./auth-reducer"
 
 
 let reducers = combineReducers({
@@ -17,8 +18,8 @@ let reducers = combineReducers({
     app: appReducer
 })
 
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-const store = createStore(reducers, composeEnhancers(applyMiddleware(thunkMiddleware)));
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
+const store = createStore(reducers, composeEnhancers(applyMiddleware(thunkMiddleware)))
 
 window.__store__ = store
 
