@@ -1,13 +1,13 @@
 import {usersAPI} from "../api/api";
-import {updateObjectInArray} from "../utils/object-helpers";
-// like in auth reducer
-const FOLLOW = 'FOLLOW'
-const UNFOLLOW = 'UNFOLLOW'
-const SET_USERS = 'SET_USERS'
-const SET_CURRENT_PAGE = 'SET_CURRENT_PAGE'
-const SET_TOTAL_USERS_COUNT = 'SET_TOTAL_USERS_COUNT'
-const TOGGLE_IS_FETCHING = 'TOGGLE_IS_FETCHING'
-const TOGGLE_IS_FOLLOWING_IN_PROGRESS = 'TOGGLE_IS_FOLLOWING_IN_PROGRESS'
+import {updateObjectInArray} from "../utils/object-helpers"
+
+const FOLLOW = '/users/FOLLOW'
+const UNFOLLOW = '/users/UNFOLLOW'
+const SET_USERS = '/users/SET_USERS'
+const SET_CURRENT_PAGE = '/users/SET_CURRENT_PAGE'
+const SET_TOTAL_USERS_COUNT = '/users/SET_TOTAL_USERS_COUNT'
+const TOGGLE_IS_FETCHING = '/users/TOGGLE_IS_FETCHING'
+const TOGGLE_IS_FOLLOWING_IN_PROGRESS = '/users/TOGGLE_IS_FOLLOWING_IN_PROGRESS'
 
 let initialState = {
     users: [ ],
